# frontend/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AI Chat API",
  "description": "Backend API for the AI companion/chat application: authentication, AI characters, conversations, messages, memories, and streamed AI replies (SSE). All endpoints except `/auth/register` and `/auth/login` require a Bearer token.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/me": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Get the authenticated user's profile",
    "operationId": "get_me_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserRead"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "users"
    ],
    "summary": "Update the authenticated user's profile",
    "description": "Update the display name and/or the default AI model.",
    "operationId": "update_me_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/conversations": {
   "get": {
    "tags": [
     "conversations"
    ],
    "summary": "List the user's conversations",
    "operationId": "list_conversations_conversations_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ConversationRead"
         },
         "title": "Response List Conversations Conversations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "conversations"
    ],
    "summary": "Create a new conversation",
    "description": "Starts a conversation with an AI character. The model defaults to the user's default model, then to the first active model.",
    "operationId": "create_conversation_conversations_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConversationCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/conversations/{conversation_id}": {
   "get": {
    "tags": [
     "conversations"
    ],
    "summary": "Get one of the user's conversations",
    "operationId": "get_conversation_conversations__conversation_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Conversation Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "conversations"
    ],
    "summary": "Rename a conversation",
    "operationId": "update_conversation_conversations__conversation_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Conversation Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConversationUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "conversations"
    ],
    "summary": "Delete a conversation (soft delete)",
    "description": "Marks the conversation DELETED; the row and its messages are preserved.",
    "operationId": "delete_conversation_conversations__conversation_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Conversation Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "CharacterBrief": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "CharacterBrief"
   },
...
   "ConversationCreate": {
    "properties": {
     "character_id": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Character Id"
     },
     "model_id": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Id",
      "description": "Defaults to the user's default model"
     },
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "type": "object",
    "required": [
     "character_id"
    ],
    "title": "ConversationCreate",
    "description": "Body for POST /conversations."
   },
   "ConversationRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "character": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CharacterBrief"
       },
       {
        "type": "null"
       }
      ]
     },
     "model": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ModelBrief"
       },
       {
        "type": "null"
       }
      ]
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "last_message_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Message At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "character",
     "model",
     "title",
     "status",
     "created_at",
     "updated_at",
     "last_message_at"
    ],
    "title": "ConversationRead"
   },
   "ConversationUpdate": {
    "properties": {
     "title": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Title"
     }
    },
    "type": "object",
    "required": [
     "title"
    ],
    "title": "ConversationUpdate",
    "description": "Body for PATCH /conversations/{id}."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ModelBrief": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "model_name": {
      "type": "string",
      "title": "Model Name"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "model_name",
     "display_name"
    ],
    "title": "ModelBrief"
   },
...
   "UserRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "default_model_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Model Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "username",
     "display_name",
     "role",
     "status",
     "default_model_id",
     "created_at",
     "last_login_at"
    ],
    "title": "UserRead",
    "description": "Public user profile. Never contains the password hash."
   },
   "UserUpdate": {
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "default_model_id": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Model Id"
     }
    },
    "type": "object",
    "title": "UserUpdate",
    "description": "Body for PATCH /me — at least one field must be provided."
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "JWT access token from POST /auth/login",
    "scheme": "bearer"
   }
  }
 }
}